import refactors - kamangir/bolt#746
