+++ notebooks_and_scripts/__init__.py
@@ -4,7 +4,7 @@ ICON = "📜"
 
 DESCRIPTION = f"{ICON} notebooks and scripts for ai experiments and aws batch jobs."
 
-VERSION = "4.693.1"
+VERSION = "4.694.1"
 
 REPO_NAME = "notebooks-and-scripts"
